+++ NEWS.md
@@ -1,5 +1,7 @@
 # pdqr (development version)
 
+* `summ_spread()` gets new method "range" with corresponding new `summ_range()`
+function. It returns a range length of "x" values within region of positive probability.
 * `form_retype()` gets new method "value" that preserves ratio of d-function values at certain "x" points.
 
 # pdqr 0.1.0